+++ packages/shared/src/index.ts
@@ -1,3 +1,5 @@
 export * from './types'
 export * from './schemas'
 export * from './api-client'
+export * from './geo'
+export * from './gpx'